Moved SidebarCount component to src/components/SidebarCount
