+++ src/components/SidebarCount/__tests__/SidebarCount.test.js
@@ -2,7 +2,7 @@ import React from "react";
 import { Provider } from "react-redux";
 import configureStore from "redux-mock-store";
 import thunk from "redux-thunk";
-import SidebarCount from "../SidebarCount";
+import SidebarCount from "../index";
 import renderer from "react-test-renderer";
 import "jest-styled-components";
 import { reduxInitialState } from "../../../constants/CommonTestData";
